api,storage,sqlite: volume ID type consistency

+++ host/storage/volume.go
@@ -49,7 +49,7 @@ type (
 
 	// A Volume stores and retrieves sector data
 	Volume struct {
-		ID           int    `json:"ID"`
+		ID           int64  `json:"ID"`
 		LocalPath    string `json:"localPath"`
 		UsedSectors  uint64 `json:"usedSectors"`
 		TotalSectors uint64 `json:"totalSectors"`
